- changed the amount field data type

+++ src/main/java/com/cba/core/wiremeweb/dto/TransactionCoreResponseDto.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -24,7 +25,7 @@ public class TransactionCoreResponseDto implements Serializable {
     private String terminalId;
     private Integer traceNo;
     private Integer invoiceNo;
-    private int amount;
+    private BigDecimal amount;
     private String currency;
     private Integer batchNo;
     private String pan;
@@ -33,7 +34,7 @@ public class TransactionCoreResponseDto implements Serializable {
     private String rrn;
     private String authCode;
     private String signData;
-    private Integer tipAmount;
+    private BigDecimal tipAmount;
     private String entryMode;
     private String dccCurrency;
     private Integer dccTranAmount;
